only delete user car when I am driving # Everything below will be removed. # Please enter the commit message for your changes. Lines starting # with '#' will be ignored, and an empty message aborts the commit. # # On branch dev # Your branch is up-to-date with 'origin/dev'. # Changes to be committed: # # 	modified:   api/app/routes/events.php # diff --git a/api/app/routes/events.php b/api/app/routes/events.php index 07cbfcc5..0580913c 100644 --- a/api/app/routes/events.php +++ b/api/app/routes/events.php @@ -645,16 +645,19 @@ $app->group('/events', function () {          $can_drive = (bool) $formData['can_drive'];          $drivers_req = (bool) $event->drivers_required;        	if($drivers_req && $user_type == 'Mentor') { +          $car = FrcPortal\EventCar::find($reqUpdate->car_id);            if($can_drive) {              $eventCarUpdate = FrcPortal\EventCar::updateOrCreate(['event_id' => $event_id, 'user_id' => $user_id], ['car_space' => $formData['event_cars']['car_space']]);              $reqUpdate->can_drive = true;              $reqUpdate->car_id = $eventCarUpdate->car_id;              $reqUpdate->save();            } else { -            $eventCarUpdate = FrcPortal\EventCar::where('event_id',$event_id)->where('user_id',$user_id)->delete(); -            $reqUpdate->can_drive = false; -            $reqUpdate->car_id = null; -            $reqUpdate->save(); +            if(!is_null($car) && $car->user_id == $user_id) { +              $eventCarUpdate = FrcPortal\EventCar::where('event_id',$event_id)->where('user_id',$user_id)->delete(); +              $reqUpdate->can_drive = false; +              $reqUpdate->car_id = null; +              $reqUpdate->save(); +            }            }          }          $room_required = (bool) $event->room_required;

+++ api/app/routes/events.php
@@ -645,16 +645,19 @@ $app->group('/events', function () {
         $can_drive = (bool) $formData['can_drive'];
         $drivers_req = (bool) $event->drivers_required;
       	if($drivers_req && $user_type == 'Mentor') {
+          $car = FrcPortal\EventCar::find($reqUpdate->car_id);
           if($can_drive) {
             $eventCarUpdate = FrcPortal\EventCar::updateOrCreate(['event_id' => $event_id, 'user_id' => $user_id], ['car_space' => $formData['event_cars']['car_space']]);
             $reqUpdate->can_drive = true;
             $reqUpdate->car_id = $eventCarUpdate->car_id;
             $reqUpdate->save();
           } else {
-            $eventCarUpdate = FrcPortal\EventCar::where('event_id',$event_id)->where('user_id',$user_id)->delete();
-            $reqUpdate->can_drive = false;
-            $reqUpdate->car_id = null;
-            $reqUpdate->save();
+            if(!is_null($car) && $car->user_id == $user_id) {
+              $eventCarUpdate = FrcPortal\EventCar::where('event_id',$event_id)->where('user_id',$user_id)->delete();
+              $reqUpdate->can_drive = false;
+              $reqUpdate->car_id = null;
+              $reqUpdate->save();
+            }
           }
         }
         $room_required = (bool) $event->room_required;
